Moves team availability code to ee
