+++ ee/components/team/availability/TeamAvailabilityModal.tsx
